requirements-all.txt

backcall==0.1.0
bert-tensorflow==1.0.1
bleach==3.1.0
blis==0.4.1
bresenham==0.2.1
catalogue==1.0.0
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
cymem==2.0.3
decorator==4.4.1
defusedxml==0.6.0
dill==0.3.1.1
docopt==0.6.2
docutils==0.16
easydict==1.9
EMD-signal==0.2.8
en-core-web-sm==2.2.5
entrypoints==0.3
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Login==0.5.0
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.18.2
gast==0.2.2
geographiclib==1.50
geopy==2.0.0
gevent==20.9.0
gevent-websocket==0.10.1
git-filter-repo==2.29.0
googledrivedownloader==0.4
greenlet==0.4.17
grpcio==1.21.1
guppy3==3.0.10.post1
h5py==2.9.0
idna==2.8
imagesize==1.2.0
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.16.0
jieba==0.42.1
Jinja==22.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
kaggle==1.5.6
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
llvmlite==0.33.0
Mako==1.1.3
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib3.1.1
memory-profiler==0.57.0
MetPy==0.12.1
mistune==0.8.4
mock==3.0.5
mpmath==1.1.0
multiprocess==0.70.9
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
nose==1.3.7
notebook==6.0.3
numba==0.50.1
numpy==1.17.4
numpydoc==0.9.2
packaging==20.1
pandas==0.25.3
pandocfilters==1.4.2
parso==0.6.1
pathos==0.2.5
patsy==0.5.1
pickleshare==0.7.5
Pillow==7.0.0
Pint==0.14
pipreqs==0.4.10
plac==1.1.3
ply==3.11
plyfile==0.7.2
pooch==1.1.1
pox==0.2.7
ppft==1.6.6.1
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.7.1
psutil==5.7.2
pycparser==2.20
pycuda==2019.1.2
Pygments==2.5.2
PyMySQL==0.10.1
pyparsing==2.4.0
pyrsistent==0.15.7
pytest-runner==5.2
python-dateutil==2.8.0
python-engineio==3.13.2
python-slugify==4.0.0
python-socketio==4.6.0
pytools==2020.4
pytz==2019.3
PyWavelets==1.1.1
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==18.1.1
qtconsole==4.6.0
rdflib==5.0.0
requests==2.22.0
scikit-learn==0.24.2
scikit-opt==0.5.5
scipy==1.4.1
Send2Trash==1.5.0
setuptools==50.3.0
six==1.12.0
smop==0.41
snowballstemmer==2.0.0
spacy==2.2.3
Sphinx==2.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.20
srsly==1.0.1
statsmodels==0.11.1
sympy==1.5.1
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
tflib==0.1.0
thinc==7.3.1
threadpoolctl==2.1.0
torch-geometric==1.6.0
torch-scatter==2.0.5
torch-sparse==0.6.6
torchvision==0.6.1
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.3
urllib3==1.24.3
wasabi==0.6.0
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.15.4
wheel==0.33.4
widgetsnbextension==3.5.1
WTForms==2.2.1
WunderWeather==1.1.0
xarray==0.16.0
xlrd==1.2.0
xlwt==1.3.0
xx-ent-wiki-sm==2.2.0
yarg==0.1.9
zipp==2.2.0
zope.event==4.5.0
zope.interface==5.1.2
